move tmp helper func into packer-plugin-sdk
