+++ builder/googlecompute/step_start_tunnel.go
@@ -23,7 +23,7 @@ import (
 	"github.com/hashicorp/packer/helper/communicator"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 // StepStartTunnel represents a Packer build step that launches an IAP tunnel
