+++ builder/lxc/communicator.go
@@ -13,7 +13,7 @@ import (
 	"syscall"
 
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 type LxcAttachCommunicator struct {
