+++ builder/osc/chroot/communicator.go
@@ -14,7 +14,7 @@ import (
 	"syscall"
 
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 // Communicator is a special communicator that works by executing
