+++ builder/virtualbox/common/step_attach_floppy.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 // This step attaches the ISO to the virtual machine.
