+++ builder/vmware/vmx/step_clone_vmx.go
@@ -11,7 +11,7 @@ import (
 	vmwcommon "github.com/hashicorp/packer/builder/vmware/common"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 // StepCloneVMX takes a VMX file and clones the VM into the output directory.
