+++ common/adapter/scp.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 const (
