+++ common/commonsteps/step_create_cdrom.go
@@ -14,7 +14,7 @@ import (
 	"github.com/hashicorp/packer/common/shell-local/localexec"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 // StepCreateCD will create a CD disk with the given files.
