+++ packer-plugin-sdk/sdk-internals/communicator/ssh/communicator.go
@@ -17,7 +17,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 	"github.com/pkg/sftp"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
