+++ post-processor/vagrant/util.go
@@ -12,7 +12,7 @@ import (
 	"runtime"
 
 	"github.com/hashicorp/packer/packer"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 	"github.com/klauspost/pgzip"
 )
