+++ provisioner/ansible-local/provisioner.go
@@ -15,7 +15,7 @@ import (
 	"github.com/hashicorp/packer/helper/config"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/template/interpolate"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 const DefaultStagingDir = "/tmp/packer-provisioner-ansible-local"
