+++ provisioner/windows-shell/provisioner.go
@@ -22,7 +22,7 @@ import (
 	"github.com/hashicorp/packer/helper/config"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/template/interpolate"
-	"github.com/hashicorp/packer/packer/tmp"
+	"github.com/hashicorp/packer/packer-plugin-sdk/tmp"
 )
 
 //FIXME query remote host or use %SYSTEMROOT%, %TEMP% and more creative filename
